tidy: drying up seo code
