+++ src/app/(with-color-blur)/about/page.tsx
@@ -3,6 +3,7 @@ import { Balancer } from 'react-wrap-balancer';
 import { getAboutPage } from '~/api/about';
 import { heading, text } from '~/components/ui/text';
 import { PATHS } from '~/constants/paths.constants';
+import { getMetadataFromSeo } from '~/utils/seo';
 
 export const revalidate = 300; // 5 minutes
 
@@ -11,22 +12,7 @@ export async function generateMetadata(): Promise<Metadata> {
 		const data = await getAboutPage();
 		const { seo } = data.aboutPage ?? {};
 
-		const title = seo?.title ?? '';
-		const description = seo?.description ?? '';
-		const index = Boolean(seo?.index);
-
-		return {
-			title,
-			description,
-			robots: index ? 'index, follow' : 'noindex, nofollow',
-			openGraph: {
-				title,
-				description,
-				type: 'website',
-				locale: 'en_US',
-				url: `${PATHS.absolute}${PATHS.about}`,
-			},
-		};
+		return getMetadataFromSeo(`${PATHS.absolute}${PATHS.about}`, seo);
 	} catch (_error) {
 		console.error('Error generating metadata for about page');
