+++ src/app/(with-color-blur)/schedule/page.tsx
@@ -4,29 +4,14 @@ import { getSchedulePage } from '~/api/schedule';
 import { Container } from '~/components/container';
 import { heading } from '~/components/ui/text';
 import { PATHS } from '~/constants/paths.constants';
+import { getMetadataFromSeo } from '~/utils/seo';
 import styles from './page.module.css';
 
 export async function generateMetadata(): Promise<Metadata> {
 	try {
 		const data = await getSchedulePage();
 		const { seo } = data.schedulePage ?? {};
-
-		const title = seo?.title ?? '';
-		const description = seo?.description ?? '';
-		const index = Boolean(seo?.index);
-
-		return {
-			title,
-			description,
-			robots: index ? 'index, follow' : 'noindex, nofollow',
-			openGraph: {
-				title,
-				description,
-				type: 'website',
-				locale: 'en_US',
-				url: `${PATHS.absolute}${PATHS.schedule}`,
-			},
-		};
+		return getMetadataFromSeo(`${PATHS.absolute}${PATHS.schedule}`, seo);
 	} catch (_error) {
 		console.error('Error generating metadata for schedule page');
